refactored ProfilePage  component
