+++ modules/ProfilePage/index.ts
@@ -1,1 +1,1 @@
-export { default as ProfilePage } from './pages/profile-page';
+export { default as ProfilePage } from './pages/ProfilePage';
