+++ modules/ProfilePage/pages/ProfilePage.module.css
@@ -1,0 +1,7 @@
+.loader {
+  @apply bg-grey-bg shadow-soft p-4 rounded-md flex justify-start items-center mb-6;
+}
+
+.profile-user {
+  @apply w-full flex items-center justify-center;
+}
