--- modules/ProfilePage/pages/profile-page.module.css
@@ -1,15 +1,0 @@
-.soon-parent {
-  @apply absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2;
-}
-
-.soon-text {
-  @apply text-lg sm:text-2xl md:text-3xl font-bold text-dark-grey;
-}
-
-.loader {
-  @apply bg-grey-bg shadow-soft p-4 rounded-md flex justify-start items-center mb-6;
-}
-
-.profile-user {
-  @apply w-full flex items-center justify-center;
-}
